+++ src/oss/uploadService.ts
@@ -64,6 +64,7 @@ const upload = async ({ type = 'img', file, source = '', datetime = '', storaget
                 type: file.type,
                 date: date,
                 datetime: datetime,
+                storage: storagetype,
                 file
               })
             }
@@ -102,6 +103,7 @@ const upload = async ({ type = 'img', file, source = '', datetime = '', storaget
             type: file.type,
             date: date,
             datetime: datetime,
+            storage: storagetype,
             file
           })
         }
@@ -155,6 +157,7 @@ const upload = async ({ type = 'img', file, source = '', datetime = '', storaget
             type: file.type,
             date: date,
             datetime: datetime,
+            storage: storagetype,
             file
           })
         }