+++ docs/introduction/changelog.md
@@ -1,17 +1,28 @@
 # 更新日志
 
+## v1.0.1
+
+- feat: 优化云储存引用 ossclient 方法 支持上传后返回 storage
+- fix(#66015): app 定位改成 30 秒超时
+- fix(#64389): 兼容 web 引擎开发者模式设置 isdebugger
+- fix(#66015): web 引擎端 isdebugger 模式不开启 vconsole
+
 ## v1.0.0
+
 - feat(#22295): 玄瞳域名修改
 
 ## v0.0.38
+
 - feat: 调试模式下 axios 加 debug: true 请求头
 - feat: 支持 isdebugger 和 isvirtuallocation 方法
-- feat: 存储支持minio
+- feat: 存储支持 minio
 
 ## v0.0.37（取消发版）
+
 - feat: 调试模式下 axios 加 debug: true 请求头
 - feat: 支持 isdebugger 和 isvirtuallocation 方法
-- feat: 存储支持minio
+- feat: 存储支持 minio
 
 ## v0.0.36
+
 - 初始化